move the API from vm_version_x86 to vm_version_bsd_x86
